Confirmation delete dialog

See merge request los-ng/fe!2069

+++ src/main/webapp/app/entities/collateral-appraisal/collateral/collateral-appraisal-process-detail-mesin.component.ts
@@ -11,6 +11,7 @@ import { STATUS } from 'app/shared/constants/status.constants';
 import { ICollateralAppraisal } from '../collateral-appraisal.model';
 import { AccountService } from 'app/core/auth/account.service';
 import { Account } from 'app/core/auth/account.model';
+import { ConfirmDialogComponent } from 'app/layouts/miscellaneous/confirm-dialog.component';
 @Component({
   selector: 'jhi-collateral-appraisal-process-detail-mesin',
   templateUrl: './collateral-appraisal-process-detail-mesin.component.html',
@@ -86,11 +87,29 @@ export class CollateralAppraisalDetailProcessMesinComponent implements OnChanges
     });
   }
 
+  // Delete Confirmation
   public deleteMchine(element): void {
-    this.collateralPropertyService.delete(element.id).subscribe(() => {
-      this.getData();
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '40vw',
+      data: {
+        title: 'Delete Machine',
+        message: 'Are you sure to delete ' + element.machineName + ' this data?',
+      },
+    });
+    dialogRef.afterClosed().subscribe(res => {
+      if (res) {
+        this.collateralPropertyService.delete(element.id).subscribe(() => {
+          this.getData();
+        });
+      }
     });
   }
+
+  // public deleteMchine(element): void {
+  //   this.collateralPropertyService.delete(element.id).subscribe(() => {
+  //     this.getData();
+  //   });
+  // }
   hideordisable() {
     if (this.collateralAppraisal.statusId === STATUS.APPROVE || this.collateralAppraisal.statusId === STATUS.COMPLETE) {
       return true;
